refactor: update utility import paths and add cn utility function
