+++ src/shared/components/ui/button.tsx
@@ -1,5 +1,5 @@
 import { Slot } from '@radix-ui/react-slot'
-import { cn } from '@utils/utils'
+import { cn } from '@utils/cn'
 import { cva, type VariantProps } from 'class-variance-authority'
 import * as React from 'react'
