+++ src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './utils'
+export * from './cn'
